backend/requirements.txt
# ReceptAI Backend Dependencies

# FastAPI framework
fastapi==0.110.0
uvicorn[standard]==0.27.0

# ML/AI dependencies
# YOLOv10 требует ultralytics >= 8.2.0
ultralytics==8.2.2

# PyTorch — фиксированные версии
torch==2.2.1
torchvision==0.17.1

# OpenCV headless
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# LLM for recipe generation
openai>=1.10.0

# API utilities
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database — PostgreSQL
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Auth — JWT + bcrypt
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# HTTP client
httpx>=0.26.0
aiofiles>=23.2.0

# Development
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
